requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
Flask-Migrate==4.0.5
email-validator==2.0.0
Pillow==10.4.0
python-dateutil==2.8.2
psycopg2-binary==2.9.9
PyPDF2==3.0.1
pdfplumber==0.11.0
matplotlib==3.8.2
scikit-learn==1.4.0
numpy>=1.26.0